# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OfficialUM1 API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  description: string;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  category: string;
  /** @nullable */
  platform?: string | null;
  /** @nullable */
  imageUrl?: string | null;
  stock: number;
  inStock: boolean;
  featured: boolean;
  /** @nullable */
  rating?: number | null;
  reviewCount: number;
  /** @nullable */
  badge?: string | null;
  tags: string[];
  createdAt: string;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  offset: number;
  limit: number;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  productCount: number;
}

export interface CartItem {
  productId: number;
  name: string;
  price: number;
  /** @nullable */
  imageUrl?: string | null;
  quantity: number;
}

export interface Cart {
  items: CartItem[];
  subtotal: number;
  total: number;
  itemCount: number;
}

export interface AddToCartBody {
  productId: number;
  quantity: number;
}

export interface Order {
  id: number;
  status: string;
  total: number;
  items: CartItem[];
  paymentMethod: string;
  createdAt: string;
  /** @nullable */
  deliveryToken?: string | null;
}

export interface CheckoutBody {
  paymentMethod: string;
  email: string;
  name: string;
  /** @nullable */
  couponCode?: string | null;
}

export interface CheckoutResponse {
  orderId: number;
  /** @nullable */
  deliveryToken?: string | null;
  message: string;
}

export interface RecentOrder {
  productName: string;
  /** @nullable */
  platform?: string | null;
  timeAgo: string;
}

export interface RegisterBody {
  email: string;
  password: string;
  name: string;
  /** @nullable */
  referralCode?: string | null;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  role: string;
  walletBalance: number;
  referralCode: string;
  /** @nullable */
  membershipTier?: string | null;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
  message: string;
}

export interface DashboardData {
  user: User;
  recentOrders: Order[];
  walletBalance: number;
  totalOrders: number;
  totalSpent: number;
  referralEarnings: number;
  openTickets: number;
}

export interface BlogPost {
  id: number;
  title: string;
  slug: string;
  excerpt: string;
  content: string;
  /** @nullable */
  imageUrl?: string | null;
  category: string;
  author: string;
  readTime: number;
  createdAt: string;
}

export interface Review {
  id: number;
  /** @nullable */
  productId?: number | null;
  /** @nullable */
  productName?: string | null;
  authorName: string;
  rating: number;
  comment: string;
  verified: boolean;
  createdAt: string;
}

export interface CreateReviewBody {
  /** @nullable */
  productId?: number | null;
  rating: number;
  comment: string;
  authorName: string;
}

export interface ReviewStats {
  averageRating: number;
  totalReviews: number;
  fiveStar: number;
  fourStar: number;
  threeStar: number;
  twoStar: number;
  oneStar: number;
}

export interface Service {
  id: number;
  title: string;
  description: string;
  icon: string;
  /** @nullable */
  price?: number | null;
  priceLabel: string;
  features: string[];
  popular: boolean;
}

export interface PublicStats {
  totalOrders: number;
  happyClients: number;
  productsAvailable: number;
  averageRating: number;
  totalReviews: number;
}

export interface ContactBody {
  name: string;
  email: string;
  subject: string;
  message: string;
  /** @nullable */
  service?: string | null;
}

export interface NewsletterBody {
  email: string;
  /** @nullable */
  name?: string | null;
}

export interface TicketReply {
  id: number;
  author: string;
  isStaff: boolean;
  message: string;
  createdAt: string;
}

export interface Ticket {
  id: number;
  subject: string;
  status: string;
  priority: string;
  message: string;
  replies: TicketReply[];
  createdAt: string;
}

export interface CreateTicketBody {
  subject: string;
  message: string;
  priority: string;
  /** @nullable */
  orderId?: number | null;
}

export interface TicketReplyBody {
  message: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export type ListProductsParams = {
  /**
   * @nullable
   */
  category?: string | null;
  /**
   * @nullable
   */
  platform?: string | null;
  /**
   * @nullable
   */
  search?: string | null;
  /**
   * @nullable
   */
  minPrice?: number | null;
  /**
   * @nullable
   */
  maxPrice?: number | null;
  /**
   * @nullable
   */
  inStock?: boolean | null;
  /**
   * @nullable
   */
  featured?: boolean | null;
  /**
   * @nullable
   */
  limit?: number | null;
  /**
   * @nullable
   */
  offset?: number | null;
};

export type ListBlogsParams = {
  /**
   * @nullable
   */
  limit?: number | null;
  /**
   * @nullable
   */
  offset?: number | null;
  /**
   * @nullable
   */
  category?: string | null;
};

export type ListReviewsParams = {
  /**
   * @nullable
   */
  productId?: number | null;
  /**
   * @nullable
   */
  limit?: number | null;
};