# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: OfficialUM1 API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cart/add:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
  /cart/remove/{productId}:
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
...
  /auth/register:
    post:
      operationId: register
      tags: [users]
      summary: Register a new buyer account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [users]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
...
  /blogs:
    get:
      operationId: listBlogs
      tags: [blogs]
      summary: List blog posts
      parameters:
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
        - name: offset
          in: query
          schema:
            type: ["integer", "null"]
        - name: category
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Blog posts list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
  /blogs/{slug}:
    get:
      operationId: getBlog
      tags: [blogs]
      summary: Get blog post by slug
      parameters:
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"
        "404":
          description: Not found

  # ── REVIEWS ───────────────────────────────────────────────────
...
  /newsletter/subscribe:
    post:
      operationId: subscribeNewsletter
      tags: [newsletter]
      summary: Subscribe to newsletter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsletterBody"
      responses:
        "201":
          description: Subscribed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── TICKETS ───────────────────────────────────────────────────
...
components:
  schemas:
...
    CartItem:
      type: object
      properties:
        productId:
          type: integer
        name:
          type: string
        price:
          type: number
        imageUrl:
          type: ["string", "null"]
        quantity:
          type: integer
      required: [productId, name, price, quantity]
    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        subtotal:
          type: number
        total:
          type: number
        itemCount:
          type: integer
      required: [items, subtotal, total, itemCount]
    AddToCartBody:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required: [productId, quantity]

    # ── ORDERS ────────────────────────────────────────────────
...
    RegisterBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        name:
          type: string
        referralCode:
          type: ["string", "null"]
      required: [email, password, name]
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
        walletBalance:
          type: number
        referralCode:
          type: string
        membershipTier:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, email, name, role, walletBalance, referralCode, createdAt]
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
        message:
          type: string
      required: [user, token, message]
...
    BlogPost:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        excerpt:
          type: string
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        category:
          type: string
        author:
          type: string
        readTime:
          type: integer
        createdAt:
          type: string
          format: date-time
      required: [id, title, slug, excerpt, content, category, author, readTime, createdAt]

    # ── REVIEWS ───────────────────────────────────────────────
...
    NewsletterBody:
      type: object
      properties:
        email:
          type: string
        name:
          type: ["string", "null"]
      required: [email]

    # ── TICKETS ───────────────────────────────────────────────
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success, message]